# course-projects/8linkedin-2back/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LinkedIn Clone API",
    "description": "API documentation for the LinkedIn Clone project, managing authentication, users, connections, conversations, and applications.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3072"
    }
  ],
...
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets all users.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Creates a new user.",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "headline": {
                    "example": "any"
                  },
                  "about": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "experience": {
                    "example": "any"
                  },
                  "education": {
                    "example": "any"
                  },
                  "skills": {
                    "example": "any"
                  },
                  "isAdmin": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets a user by ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Updates a user by ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  },
                  "headline": {
                    "example": "any"
                  },
                  "about": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "experience": {
                    "example": "any"
                  },
                  "education": {
                    "example": "any"
                  },
                  "skills": {
                    "example": "any"
                  },
                  "isAdmin": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Deletes a user by ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/posts/": {
      "get": {
        "tags": [
          "Posts"
        ],
        "description": "Retrieve all posts",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "description": "Create a new post",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "Gets all comments for a specific post.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "description": "Creates a new comment for a specific post.",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "post": {
                    "example": "any"
                  },
                  "text": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "description": "Retrieve all jobs",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "description": "Create a new job",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "requirements": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/": {
      "get": {
        "tags": [
          "Applications"
        ],
        "description": "Gets all applications for a specific job.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Applications"
        ],
        "description": "Creates a new application for a specific job.",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "job": {
                    "example": "any"
                  },
                  "resume": {
                    "example": "any"
                  },
                  "coverLetter": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/conversations/": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "description": "Retrieve all conversations",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Conversations"
        ],
        "description": "Create a new conversation with an initial message",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "participants": {
                    "example": "any"
                  },
                  "text": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/connections/": {
      "get": {
        "tags": [
          "Connections"
        ],
        "description": "Retrieve all connections",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Connections"
        ],
        "description": "Create a new connection",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiver": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}